pillstore/requirements.txt
aiofiles==25.1.0 ; python_version >= "3.12" and python_version < "4.0"
aiohappyeyeballs==2.6.1 ; python_version >= "3.12" and python_version < "4.0"
aiohttp==3.13.3 ; python_version >= "3.12" and python_version < "4.0"
aiosignal==1.4.0 ; python_version >= "3.12" and python_version < "4.0"
aiosqlite==0.22.1 ; python_version >= "3.12" and python_version < "4.0"
alembic==1.17.2 ; python_version >= "3.12" and python_version < "4.0"
annotated-doc==0.0.4 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.12.0 ; python_version >= "3.12" and python_version < "4.0"
asyncpg==0.31.0 ; python_version >= "3.12" and python_version < "4.0"
attrs==25.4.0 ; python_version >= "3.12" and python_version < "4.0"
bcrypt==4.0.1 ; python_version >= "3.12" and python_version < "4.0"
beautifulsoup4==4.14.3 ; python_version >= "3.12" and python_version < "4.0"
bs4==0.0.2 ; python_version >= "3.12" and python_version < "4.0"
certifi==2026.1.4 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.4 ; python_version >= "3.12" and python_version < "4.0"
chromium==0.0.0 ; python_version >= "3.12" and python_version < "4.0"
click==8.3.1 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
dnspython==2.8.0 ; python_version >= "3.12" and python_version < "4.0"
email-validator==2.3.0 ; python_version >= "3.12" and python_version < "4.0"
fastapi==0.121.3 ; python_version >= "3.12" and python_version < "4.0"
frozenlist==1.8.0 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.3.0 ; python_version >= "3.12" and python_version < "4.0"
h11==0.16.0 ; python_version >= "3.12" and python_version < "4.0"
httptools==0.7.1 ; python_version >= "3.12" and python_version < "4.0"
idna==3.11 ; python_version >= "3.12" and python_version < "4.0"
iniconfig==2.3.0 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.6 ; python_version >= "3.12" and python_version < "4.0"
jose==1.0.0 ; python_version >= "3.12" and python_version < "4.0"
lxml==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
mako==1.3.10 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==3.0.3 ; python_version >= "3.12" and python_version < "4.0"
multidict==6.7.0 ; python_version >= "3.12" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.12" and python_version < "4.0"
passlib==1.7.4 ; python_version >= "3.12" and python_version < "4.0"
playwright==1.57.0 ; python_version >= "3.12" and python_version < "4.0"
pluggy==1.6.0 ; python_version >= "3.12" and python_version < "4.0"
propcache==0.4.1 ; python_version >= "3.12" and python_version < "4.0"
pycryptodome ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.41.5 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.12.5 ; python_version >= "3.12" and python_version < "4.0"
pydantic[email]==2.12.5 ; python_version >= "3.12" and python_version < "4.0"
pyee==13.0.0 ; python_version >= "3.12" and python_version < "4.0"
pygments==2.19.2 ; python_version >= "3.12" and python_version < "4.0"
pyjwt==2.10.1 ; python_version >= "3.12" and python_version < "4.0"
pytest-asyncio==1.3.0 ; python_version >= "3.12" and python_version < "4.0"
pytest-mock==3.15.1 ; python_version >= "3.12" and python_version < "4.0"
pytest==9.0.2 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.2.1 ; python_version >= "3.12" and python_version < "4.0"
python-json-logger==4.0.0 ; python_version >= "3.12" and python_version < "4.0"
python-multipart==0.0.20 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.3 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.5 ; python_version >= "3.12" and python_version < "4.0"
soupsieve==2.8.1 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy==2.0.45 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.50.0 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.15.0 ; python_version >= "3.12" and python_version < "4.0"
typing-inspection==0.4.2 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.6.3 ; python_version >= "3.12" and python_version < "4.0"
uvicorn[standard]==0.38.0 ; python_version >= "3.12" and python_version < "4.0"
uvloop==0.22.1 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.12" and python_version < "4.0"
watchfiles==1.1.1 ; python_version >= "3.12" and python_version < "4.0"
websockets==15.0.1 ; python_version >= "3.12" and python_version < "4.0"
yarl==1.22.0 ; python_version >= "3.12" and python_version < "4.0"
